Cleanup platform construction.
