+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/Crypto.java
@@ -16,36 +16,25 @@
 
 package com.swirlds.platform;
 
-import static com.swirlds.logging.LogMarker.EXCEPTION;
-
 import com.swirlds.common.crypto.Signature;
 import com.swirlds.platform.crypto.CryptoStatic;
 import com.swirlds.platform.crypto.KeysAndCerts;
 import com.swirlds.platform.crypto.PlatformSigner;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
 import java.security.PublicKey;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.function.Consumer;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public class Crypto {
-    /** use this for all logging, as controlled by the optional data/log4j2.xml file */
-    private static final Logger logger = LogManager.getLogger(Crypto.class);
 
     private final KeysAndCerts keysAndCerts;
     /** a pool of threads used to verify signatures and generate keys, in parallel */
     private final ExecutorService cryptoThreadPool;
 
     /**
-     * @param keysAndCerts
-     * 		keys and certificates
-     * @param cryptoThreadPool
-     * 		the thread pool that will be used for all operations that can be done in parallel, like signing and
-     * 		verifying
+     * @param keysAndCerts     keys and certificates
+     * @param cryptoThreadPool the thread pool that will be used for all operations that can be done in parallel, like
+     *                         signing and verifying
      */
     public Crypto(final KeysAndCerts keysAndCerts, final ExecutorService cryptoThreadPool) {
         this.keysAndCerts = keysAndCerts;
@@ -53,47 +42,33 @@ public class Crypto {
     }
 
     /**
-     * Digitally sign the data with the private key. Return null if anything goes wrong (e.g., bad private
-     * key).
+     * Digitally sign the data with the private key. Return null if anything goes wrong (e.g., bad private key).
      * <p>
-     * The returned signature will be at most SIG_SIZE_BYTES bytes, which is 104 for the CNSA suite
-     * parameters.
+     * The returned signature will be at most SIG_SIZE_BYTES bytes, which is 104 for the CNSA suite parameters.
      *
-     * @param data
-     * 		the data to sign
+     * @param data the data to sign
      * @return the signature (or null if any errors)
      */
     public Signature sign(final byte[] data) {
-        try {
-            final PlatformSigner signer = new PlatformSigner(keysAndCerts);
-            return signer.sign(data);
-        } catch (NoSuchAlgorithmException | NoSuchProviderException | InvalidKeyException | RuntimeException e) {
-            logger.error(EXCEPTION.getMarker(), "ERROR in sig 3", e);
-        }
-        return null;
+        return new PlatformSigner(keysAndCerts).sign(data);
     }
 
     /**
-     * Verify the given signature for the given data. This is submitted to the thread pool so that it will
-     * be done in parallel with other signature verifications and key generation operations. This method
-     * returns a Future immediately. If the signature is valid, then a get() method on that Future will
-     * eventually return a Boolean which is true if the signature was valid. After the thread does the
-     * validation, and before it returns, it will run doLast(true) if the signature was valid, or
-     * doLast(false) if it wasn't.
+     * Verify the given signature for the given data. This is submitted to the thread pool so that it will be done in
+     * parallel with other signature verifications and key generation operations. This method returns a Future
+     * immediately. If the signature is valid, then a get() method on that Future will eventually return a Boolean which
+     * is true if the signature was valid. After the thread does the validation, and before it returns, it will run
+     * doLast(true) if the signature was valid, or doLast(false) if it wasn't.
+     * <p>
+     * This is flexible. It is OK to ignore the returned Future, and only have doLast handle the result. It is also OK
+     * to pass in (Boolean b) for doLast, and handle the result of doing a .get() on the Future. Or both mechanisms can
+     * be used.
      *
-     * This is flexible. It is OK to ignore the returned Future, and only have doLast handle the result. It
-     * is also OK to pass in (Boolean b) for doLast, and handle the result of doing a .get() on the
-     * Future. Or both mechanisms can be used.
-     *
-     * @param data
-     * 		the data that was signed
-     * @param signature
-     * 		the claimed signature of that data
-     * @param publicKey
-     * 		the claimed public key used to generate that signature
-     * @param doLast
-     * 		a function that will be run after verification, and will be passed true if the signature
-     * 		is valid. To do nothing, pass in (Boolean b)
+     * @param data      the data that was signed
+     * @param signature the claimed signature of that data
+     * @param publicKey the claimed public key used to generate that signature
+     * @param doLast    a function that will be run after verification, and will be passed true if the signature is
+     *                  valid. To do nothing, pass in (Boolean b)
      * @return validObject if the signature is valid, else returns null
      */
     public Future<Boolean> verifySignatureParallel(
