--- platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/eventflow/Failable.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (C) 2018-2023 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.swirlds.platform.test.eventflow;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-
-/**
- * A class that can have failures during the course of a unit test.
- */
-public interface Failable {
-
-    /**
-     * Indicates if there is a failure
-     *
-     * @return true is there is a failure
-     */
-    boolean hasFailure();
-
-    /**
-     * A description of the failure(s)
-     *
-     * @return the failure description
-     */
-    String getFailure();
-
-    /**
-     * Asserts there are no failures.
-     */
-    default void assertNoFailure() {
-        assertFalse(hasFailure(), getFailure());
-    }
-}
